+++ src/containers/Home/HomePage.api.js
@@ -22,6 +22,27 @@ const HomePageApi = {
 			response => response.json()
 		);
 	},
+
+	fetchShops: ({startDate=null, endDate=null}) => {
+
+		let payload = {
+			startDate,
+			endDate,
+		};
+
+		let config = {
+			method: 'post',
+			headers: {
+				'Accept': 'application/json',
+				'Content-Type': 'application/json',
+			},
+			body: JSON.stringify(payload),
+		};
+
+		return fetch(API_URL + '/home', config).then(
+			response => response.json()
+		);
+	},
 };
 
 export default HomePageApi;