+++ src/containers/Home/HomePage.reducer.js
@@ -2,7 +2,9 @@ import {
 	HOMEPAGE_SET_NEW_DATA,
 	HOMEPAGE_SET_START_DATE,
 	HOMEPAGE_SET_END_DATE,
-	FETCH_HOME_DATA_SUCCESS,
+
+	REQUEST_SHOPS,
+	RECEIVE_SHOPS,
 } from './HomePage.actions';
 
 import DateHelper from '../../common/DateHelper';
@@ -75,7 +77,11 @@ const initialState = {
 		endDate: DateHelper.getNextMonth( DateHelper.getEndOfDay( new Date() ) ),
 	},
 
-	homeData: [],
+	shops: {
+		isFetching: false,
+		lastUpdated: null,
+		items: [],
+	},
 
 };
 
@@ -106,11 +112,19 @@ const setEndDate = (state, date) => {
 	return newState;
 };
 
-const setHomeData = (state, data) => {
+const setShopsFetching = (state, isFetching) => {
 	let newState = cloneDeep(state);
-	newState.homeData = data;
+	newState.shops.isFetching = isFetching;
 	return newState;
 };
+const setShops = (state, action) => {
+	let newState = cloneDeep(state);
+	newState.shops.items = action.payload || [];
+	newState.shops.lastUpdated = action.receivedAt;
+	newState.shops.isFetching = false;
+	return newState;
+};
+
 
 export default function homePage (state = initialState, action) {
 	switch (action.type) {
@@ -120,8 +134,10 @@ export default function homePage (state = initialState, action) {
 			return setStartDate(state, action.payload);
 		case HOMEPAGE_SET_END_DATE:
 			return setEndDate(state, action.payload);
-		case FETCH_HOME_DATA_SUCCESS:
-			return setHomeData(state, action.payload);
+		case REQUEST_SHOPS:
+			return setShopsFetching(state, true);
+		case RECEIVE_SHOPS:
+			return setShops(state, action);
 		default:
 			return state;
 	}