+++ CHANGELOG.md
@@ -26,6 +26,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added API /home endpoint that receives startDate and endDate for data filtering 
 - Integrated DatePicker with data fetching
 - Added progress loaders on side menu drop downs loading
+- Added Google material icons
+- Added progress loader for home page table and no data info and icon
 
 
 ## TODO